+++ .eslintrc.json
@@ -1,9 +1,11 @@
 {
   "plugins": ["@typescript-eslint"],
   "extends": [
+    "next",
     "next/core-web-vitals",
     "plugin:jest-dom/recommended",
     "plugin:@typescript-eslint/recommended",
+    "plugin:@next/next/recommended",
     "prettier"
   ],
   "rules": {
