+++ jest.config.ts
@@ -2,7 +2,7 @@ import nextJest from "next/jest";
 // Sync object
 const createJestConfig = nextJest({
   // Provide the path to your Next.js app to load next.config.js and .env files in your test environment
-  dir: "./",
+  dir: "./src",
 });
 
 // Add any custom config to be passed to Jest
