+++ package.json
@@ -17,11 +17,11 @@
     }
   },
   "dependencies": {
-    "@types/react-table": "^7.7.10",
+    "cross-fetch": "^3.1.5",
     "drupal-jsonapi-params": "^2.0.0",
     "lodash": "^4.17.21",
-    "next": "v12.1.5-canary.7",
-    "next-drupal": "^1.2.2",
+    "next": "12.1.5-canary.7",
+    "next-drupal": "^1.3.0",
     "react": "18.0.0",
     "react-dom": "18.0.0",
     "react-table": "^7.7.0",
@@ -31,9 +31,10 @@
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.0.1",
     "@types/jest": "^27.4.1",
-    "@types/node": "^17.0.23",
+    "@types/node": "^17.0.25",
     "@types/react": "^18.0.1",
     "@types/react-dom": "^18.0.0",
+    "@types/react-table": "^7.7.10",
     "@types/testing-library__jest-dom": "^5.14.3",
     "@types/testing-library__react": "^10.2.0",
     "@types/uuid": "^8.3.4",
