+++ src/components/footer/mobile.tsx
