+++ src/components/paragraphs/table/helpers.tsx
