+++ src/types/seo.ts
