+++ tsconfig.json
@@ -14,7 +14,7 @@
     "isolatedModules": true,
     "jsx": "preserve",
     "incremental": true,
-    "baseUrl": ".",
+    "baseUrl": "./src",
     "paths": {
         "@/api/*": ["pages/api/*"],
         "@/components/*": ["components/*"],
@@ -25,6 +25,7 @@
         "@/types/*": ["types/*"],
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "**/*.js", "**/*.jsx"],
+  "types": ["node", "jest", "cypress"],
+  "include": ["next-envsrc/types/* ", "**/*.ts", "**/*.tsx", "**/*.js", "**/*.jsx"],
   "exclude": ["node_modules", "cypress"]
 }
